refactor(sdk/analyze): externalize hotspots tests to hotspots_tests.rs
hotspots.rs was 2451 lines, but ~1426 of those were the inline
`#[cfg(test)] mod tests` block — the code itself is a cohesive ~1024-line
attribution + aggregation module. Move the test block verbatim into
hotspots_tests.rs, included via `#[cfg(test)] #[path = "hotspots_tests.rs"]
mod tests;`, mirroring the existing patterns_tests.rs convention.

The source file is now 1027 lines of pure code. The module name stays `tests`
so test paths (analyze::hotspots::tests::*) are unchanged; the 24 hotspots
tests pass identically. Pure relocation — no code or test logic touched.
